requirements-test.txt
-r requirements.txt

asynctest
pytest-asyncio
pytest-cov>=2.6.0
pytest-timeout
